+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to downsample recorded near fields to speed up server-side far field projections.
 - `FieldData.apply_phase(phase)` to multiply field data by a phase.
 - Optional `phase` argument to `SimulationData.plot_field` that applies a phase to complex-valued fields.
+- Ability to window near fields for spatial filtering of far fields for both client- and server-side field projections.
 
 ### Changed
 - Indent for the json string of Tidy3D models has been changed to `None` when used internally; kept as `indent=4` for writing to `json` and `yaml` files.